@@ -1,123 +1,128 @@
 const cards = [
+  /** ======== Add your info below -- Do not modify anything outside this box ======== **/
 
-    /** ======== Add your info below -- Do not modify anything outside this box ======== **/
+  {
+    name: "Project Dock",
+    title: "JS Developer",
+    image: "PD/pd.png",
+    github: "https://github.com/ProjectDock",
+  },
+  {
+    name: "Look me cooding",
+    title: "JS Developer",
+    image: "jailsonv2/jailson.png",
+    github: "https://github.com/jailsonv2",
+  },
+  {
+    name: "Tamas Kn",
+    title: "Full Stack Developer",
+    image: "TamasKn/TamasKn.png",
+    github: "https://github.com/TamasKn",
+  },
+  {
+    name: "Saurabh kamble",
+    title: "Full Stack Developer",
+    image: "saurabh2836/game_of_throne_theme.png",
+    github: "https://github.com/saurabh2836",
+  },
 
-    {
-        name: 'Project Dock',
-        title: 'JS Developer',
-        image: 'PD/pd.png',
-        github: 'https://github.com/ProjectDock',
-    },
-    {
-        name: 'Look me cooding',
-        title: 'JS Developer',
-        image: 'jailsonv2/jailson.png',
-        github: 'https://github.com/jailsonv2',
-    },
-    {
-        name: 'Tamas Kn',
-        title: 'Full Stack Developer',
-        image: 'TamasKn/TamasKn.png',
-        github: 'https://github.com/TamasKn',
-    },
-    {
-        name: 'Saurabh kamble',
-        title: 'Full Stack Developer',
-        image: 'saurabh2836/game_of_throne_theme.png',
-        github: 'https://github.com/saurabh2836',
-    },
+  {
+    name: "Agnes Brettl",
+    title: "Front-End Developer",
+    image: "AgnesBrettl/codeTwo.png",
+    github: "https://github.com/brettl1991",
+  },
+  {
+    name: "Evgeny Pletnev",
+    title: "Web Developer",
+    image: "epletnev/image.png",
+    github: "https://github.com/epletnev",
+  },
+  {
+    name: "Aniket Pathak",
+    title: "Software Engineer",
+    image: "mynameispathak/code.png",
+    github: "https://github.com/mynameispathak",
+  },
+  {
+    name: "Rohit Kadam",
+    title: "Web Developer",
+    image: "rohitk/lookatmecode.png",
+    github: "https://github.com/rohitkadamrb",
+  },
+  {
+    name: "pixelatedjelly",
+    title: "Engineer",
+    image: "pixelatedjelly/beam_deflection_calc.png",
+    github: "https://github.com/pixelatedjelly",
+  },
+  {
+    name: "Kaunik Kamila",
+    title: "Software Developer Engineer",
+    image: "kaunik07/kaunik07.png",
+    github: "https://github.com/kaunik07",
+  },
+  {
+    name: "Rashid Makki",
+    title: "Full Stack Developer",
+    image: "rashidmakki/ss.png",
+    github: "https://github.com/rashidmakki",
+  },
+  {
+    name: "cap10kirk",
+    title: "Full Stack Developer",
+    image: "cap10kirk/cap10kirk_codes.png",
+    github: "https://github.com/Cap10Kirk",
+  },
+  {
+    name: "Meghal Donde",
+    title: "Software Developer",
+    image: "meghaldonde/meghaldonde.png",
+    github: "https://github.com/meghaldonde",
+  },
+  {
+    name: "Karim Hussain",
+    title: "Solutions Design Lead",
+    image: "ka-hn/ka-hn.PNG",
+    github: "https://github.com/ka-hn",
+  },
+  {
+    name: "Harshdeep Singh",
+    title: "Full Stack Developer",
+    image: "I-am-Harshdeep/Harsh.png",
+    github: "https://github.com/I-am-Harshdeep",
+  },
+  {
+    name: "VrajKD",
+    title: "Front End Developer",
+    image: "VrajKD/VrajKD.png",
+    github: "https://github.com/VrajKD",
+  },
+  {
+    name: "FarbrorGarbo",
+    title: "App Developer [...coding in GridConstructor]",
+    image: "FarbrorGarbo/FarbrorGarbo.png",
+    github: "https://github.com/FarbrorGarbo/GridConstructor",
+  },
+  {
+    name: "Tamas Knisz",
+    title: "Full Stack Developer",
+    image: "TamasKn/coding.png",
+    github: "https://github.com/TamasKn",
+  },
+  {
+    name: "Istvan Acs",
+    title: "Software Engineer",
+    image: "IstvanAcs/istvanacs2.jpg",
+    github: "https://github.com/St3ve89",
+  },
+  {
+    name: "Istvan Acs",
+    title: "Software Engineer",
+    image: "IstvanAcs/istvanacs3.png",
+    github: "https://github.com/St3ve89",
+  },
 
-    {
-        name: 'Agnes Brettl',
-        title: 'Front-End Developer',
-        image: 'AgnesBrettl/AgnesBrettl.png',
-        github: 'https://github.com/brettl1991',
-    },
-    {
-        name: 'Istvan Acs',
-        title: 'Software Engineer',
-        image: 'IstvanAcs/IstvanAcs.jpg',
-        github: 'https://github.com/St3ve89',
-    },
-    {
-        name: 'Evgeny Pletnev',
-        title: 'Web Developer',
-        image: 'epletnev/image.png',
-        github: 'https://github.com/epletnev',
-    },
-    {
-        name: 'Aniket Pathak',
-        title: 'Software Engineer',
-        image: 'mynameispathak/code.png',
-        github: 'https://github.com/mynameispathak',
-    },
-    {
-        name: 'Rohit Kadam',
-        title: 'Web Developer',
-        image: 'rohitk/lookatmecode.png',
-        github: 'https://github.com/rohitkadamrb',
-    },
-    {
-        name: 'pixelatedjelly',
-        title: 'Engineer',
-        image: 'pixelatedjelly/beam_deflection_calc.png',
-        github: 'https://github.com/pixelatedjelly',
-    },
-    {
-        name: 'Kaunik Kamila',
-        title: 'Software Developer Engineer',
-        image: 'kaunik07/kaunik07.png',
-        github: 'https://github.com/kaunik07',
-    },
-    {
-        name: 'Rashid Makki',
-        title: 'Full Stack Developer',
-        image: 'rashidmakki/ss.png',
-        github: 'https://github.com/rashidmakki',
-    },
-    {
-        name: 'cap10kirk',
-        title: 'Full Stack Developer',
-        image: 'cap10kirk/cap10kirk_codes.png',
-        github: 'https://github.com/Cap10Kirk',
-    },
-    {
-        name: 'Meghal Donde',
-        title: 'Software Developer',
-        image: 'meghaldonde/meghaldonde.png',
-        github: 'https://github.com/meghaldonde',
-    },
-    {
-        name: 'Karim Hussain',
-        title: 'Solutions Design Lead',
-        image: 'ka-hn/ka-hn.PNG',
-        github: 'https://github.com/ka-hn',
-    },
-    {
-        name: 'Harshdeep Singh',
-        title: 'Full Stack Developer',
-        image: 'I-am-Harshdeep/Harsh.png',
-        github: 'https://github.com/I-am-Harshdeep',
-    },
-    {
-        name: "VrajKD",
-        title: "Front End Developer",
-        image: "VrajKD/VrajKD.png",
-        github: "https://github.com/VrajKD",
-    },
-    {
-        name: "FarbrorGarbo",
-        title: "App Developer [...coding in GridConstructor]",
-        image: "FarbrorGarbo/FarbrorGarbo.png",
-        github: "https://github.com/FarbrorGarbo/GridConstructor"
-    },
-    {
-        name: 'Tamas Knisz',
-        title: 'Full Stack Developer',
-        image: 'TamasKn/coding.png',
-        github: 'https://github.com/TamasKn',
-    },
-  
   /** ======== Place your infos above ======== **/
 ];
 
